butterfly/requirements.txt
open-clip-torch==2.24.0
torch==2.3.0
torchvision==0.18.0
scikit-learn==1.4.2
transformers==4.48.0
pandas
numpy
Pillow